requirements.txt
# Required packages for local development
pandas>=1.3.0
numpy>=1.21.0
matplotlib>=3.4.0
seaborn>=0.11.0
scikit-learn>=1.0.0
mysql-connector-python>=8.0.0
python-dotenv>=0.19.0